思路初步确定

+++ Curling/strategy.h
@@ -11,7 +11,7 @@ typedef struct _GAMESTATE {
 							// if WhiteToMove = 0: First player in 1st End will shot next, 
 							//  else (WhiteToMove = 1) : Second player will shot next
 
-	float	body[16][2];	// body[n][0] : x of coordinate of n th stone
+	double	body[16][2];	// body[n][0] : x of coordinate of n th stone
 							// body[n][1] : y of coordinate of n th stone
 
 }GAMESTATE, * PGAMESTATE;
@@ -19,38 +19,38 @@ typedef struct _GAMESTATE {
 
 typedef struct  _ShotInfo
 {
-	_ShotInfo(float s, float h, float a)
+	_ShotInfo(double s, double h, double a)
 	{
 		speed = s;
 		h_x = h;
 		angle = a;
 	};
-	float speed;
-	float h_x;
-	float angle;
+	double speed;
+	double h_x;
+	double angle;
 }SHOTINFO;
 
 typedef struct _MOTIONINFO
 {
-	float x_coordinate;
-	float y_coordinate;
-	float x_velocity;
-	float y_velocity;
-	float angular_velocity;
+	double x_coordinate;
+	double y_coordinate;
+	double x_velocity;
+	double y_velocity;
+	double angular_velocity;
 }MOTIONINFO;
 
 // positions on sheet
-static const float TEE_X = (float)2.375;    // x of center of house
-static const float TEE_Y = (float)4.880;    // y of center of house
-static const float HOUSE_R = (float)1.870;  // radius of house
-static const float STONE_R = (float)0.145;  // radius of stone
+static const double TEE_X = (double)2.375;    // x of center of house
+static const double TEE_Y = (double)4.880;    // y of center of house
+static const double HOUSE_R = (double)1.870;  // radius of house
+static const double STONE_R = (double)0.145;  // radius of stone
 
 // coordinate (x, y) is in play-area if:
 //   (PLAYAREA_X_MIN < x < PLAYAREA_X_MAX && PLAYAREA_Y_MIN < y < PLAYAREA_Y_MAX)
-static const float PLAYAREA_X_MIN = (float)0.000 + STONE_R;
-static const float PLAYAREA_X_MAX = (float)0.000 + (float)4.750 - STONE_R;
-static const float PLAYAREA_Y_MIN = (float)2.650 + STONE_R;
-static const float PLAYAREA_Y_MAX = (float)2.650 + (float)8.165 + STONE_R;
+static const double PLAYAREA_X_MIN = (double)0.000 + STONE_R;
+static const double PLAYAREA_X_MAX = (double)0.000 + (double)4.750 - STONE_R;
+static const double PLAYAREA_Y_MIN = (double)2.650 + STONE_R;
+static const double PLAYAREA_Y_MAX = (double)2.650 + (double)8.165 + STONE_R;
 
 
 void getBestShot(const GAMESTATE* gs, SHOTINFO* vec_ret);
